ersah-wadasa-full/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=1.4
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
httpx
python-multipart
aiofiles
slowapi
itsdangerous